Fix acceptance probabilty and detailed balance
There were two issues. One was that the min and max radius were being
undercounted, while the other was that the change in the number of
lattice sites was not being accounted for (there is a change in the
number of ways the system can be translated up and down when changing
the lattice height).